Externalize auto cal solver outputs into a separate package
