+++ src/main/java/asl/sensor/CalProcessingServer.java
@@ -1,5 +1,9 @@
 package asl.sensor;
 
+import asl.sensor.output.CalResult;
+import asl.sensor.output.RandData;
+import asl.sensor.output.SineData;
+import asl.sensor.output.StepData;
 import java.awt.BasicStroke;
 import java.awt.Color;
 import java.awt.Font;
@@ -120,7 +124,7 @@ public class CalProcessingServer {
    * @return Data from running the experiment (plots and fit pole/zero values)
    * @throws IOException If a string does not refer to a valid accessible file
    */
-  public RandData runRand(String calFileName, String outFileName,
+  public CalResult runRand(String calFileName, String outFileName,
       String respName, boolean useEmbeddedResp, String startDate, String endDate, boolean lowFreq)
       throws IOException, SeedFormatException, CodecException {
     DateTimeFormatter dtf = DateTimeFormatter.ISO_OFFSET_DATE_TIME;
@@ -165,7 +169,7 @@ public class CalProcessingServer {
    * @return Data from running the experiment (plots and fit pole/zero values)
    * @throws IOException If a string does not refer to a valid accessible file
    */
-  public RandData populateDataAndRun(String calFileName, String outFileName,
+  public CalResult populateDataAndRun(String calFileName, String outFileName,
       String respName, boolean useEmbeddedResp, String startDate, String endDate, boolean lowFreq)
       throws IOException, SeedFormatException, CodecException {
     return
@@ -188,7 +192,7 @@ public class CalProcessingServer {
    * @return Data from running the experiment (plots and fit pole/zero values)
    * @throws IOException If a string does not refer to a valid accessible file
    */
-  public RandData runRand(String calFileNameD1, String calFileNameD2,
+  public CalResult runRand(String calFileNameD1, String calFileNameD2,
       String outFileNameD1, String outFileNameD2, String respName, boolean useEmbeddedResp,
       String startDate, String endDate, boolean lowFreq)
       throws IOException, SeedFormatException, CodecException {
@@ -358,6 +362,7 @@ public class CalProcessingServer {
       ImageIO.write(images[i], "png", out);
       pngByteArrays[i] = out.toByteArray();
     }
+
     return new StepData(pngByteArrays, fitParams, initParams);
   }
 
@@ -378,7 +383,7 @@ public class CalProcessingServer {
    * @throws IOException If a string does not refer to a valid accessible file
    */
   @SuppressWarnings("unused")
-  public RandData populateDataAndRun(String calFileNameD1, String calFileNameD2,
+  public CalResult populateDataAndRun(String calFileNameD1, String calFileNameD2,
       String outFileNameD1, String outFileNameD2, String respName, boolean useEmbeddedResp,
       String startDate, String endDate, boolean lowFreq)
       throws IOException, SeedFormatException, CodecException {
@@ -388,7 +393,7 @@ public class CalProcessingServer {
 
   }
 
-  private RandData runExpGetDataRand(DataStore dataStore, boolean isLowFrequency)
+  private CalResult runExpGetDataRand(DataStore dataStore, boolean isLowFrequency)
       throws IOException {
 
     RandomizedExperiment randomExperiment = new RandomizedExperiment();
@@ -549,157 +554,6 @@ public class CalProcessingServer {
     return new RandData(fitPoles, fitZeros, initialPoles, initialZeros, pngByteArrays,
         names, gapStarts, gapEnds);
 
-  }
-
-  @SuppressWarnings({"WeakerAccess", "unused"})
-  public abstract class CalResult {
-
-    protected Map<String, double[]> numerMap;
-    protected Map<String, byte[]> imageMap;
-
-    public CalResult() {
-      numerMap = new HashMap<>();
-      imageMap = new HashMap<>();
-    }
-
-    public Map<String, byte[]> getImageMap() {
-      return imageMap;
-    }
-
-    public Map<String, double[]> getNumerMap() {
-      return numerMap;
-    }
-  }
-
-  public class StepData extends CalResult {
-
-    // constuctor to be used with step calibrations
-    StepData(byte[][] images, double[] initParams, double[] fitParams) {
-      super();
-      double fitCorner = fitParams[0];
-      double fitDamping = fitParams[1];
-      double fitResid = fitParams[2];
-      double initCorner = initParams[0];
-      double initDamping = initParams[1];
-      double initResid = initParams[2];
-      numerMap.put("Fit_corner", new double[]{fitCorner});
-      numerMap.put("Fit_damping", new double[]{fitDamping});
-      numerMap.put("Fit_residual", new double[]{fitResid});
-      numerMap.put("Initial_corner", new double[]{initCorner});
-      numerMap.put("Initial_damping", new double[]{initDamping});
-      numerMap.put("Initial_residual", new double[]{initResid});
-      imageMap.put("Step_plot", images[0]);
-      imageMap.put("Response_amplitudes", images[1]);
-      imageMap.put("Response_phases", images[2]);
-    }
-  }
-
-  public class SineData extends CalResult {
-
-    SineData(byte[][] images, double calAmp, double outAmp, double freq, double ratio) {
-      super();
-      numerMap.put("Calibration_amplitude", new double[]{calAmp});
-      numerMap.put("Output_signal_amplitude", new double[]{outAmp});
-      numerMap.put("Estimated_signal_frequency", new double[]{freq});
-      numerMap.put("Calibration_to_output_ratio", new double[]{ratio});
-      imageMap.put("Sine_curves_plot", images[0]);
-      imageMap.put("Linearity", images[1]);
-    }
-  }
-
-  //RandData is used externally and requires Public access on getters
-  @SuppressWarnings({"WeakerAccess", "unused"})
-  public class RandData extends CalResult {
-
-    private String[] gapNameIdentifiers;
-    private Date[][] gapStarts;
-    private Date[][] gapEnds;
-
-    RandData(double[] fitPoles, double[] fitZeros, double[] initialPoles, double[] initialZeros,
-        byte[][] images,
-        String[] gapNames, Date[][] gapStartTimes, Date[][] gapEndTimes) {
-      super();
-      numerMap.put("Best_fit_poles", fitPoles);
-      numerMap.put("Best_fit_zeros", fitZeros);
-      numerMap.put("Initial_poles", initialPoles);
-      numerMap.put("Initial_zeros", initialZeros);
-      imageMap.put("Response_amplitudes", images[0]);
-      imageMap.put("Response_phases", images[1]);
-      imageMap.put("Amplitude_error", images[2]);
-      imageMap.put("Phase_error", images[3]);
-      gapNameIdentifiers = gapNames;
-      gapStarts = gapStartTimes;
-      gapEnds = gapEndTimes;
-    }
-
-    public byte[] getAmpErrorImage() {
-      return imageMap.get("Amplitude_error");
-    }
-
-    public byte[] getAmpImage() {
-      return imageMap.get("Response_amplitudes");
-    }
-
-    public double[] getFitPoles() {
-      return numerMap.get("Best_fit_poles");
-    }
-
-    public double[] getFitZeros() {
-      return numerMap.get("Best_fit_zeros");
-    }
-
-    public Date[][] getGapEndDates() {
-      return gapEnds;
-    }
-
-    public String[] getGapIdentifiers() {
-      return gapNameIdentifiers;
-    }
-
-    public String getGapInfoAsString() {
-      SimpleDateFormat sdf = new SimpleDateFormat("DD.HH:m:s");
-      sdf.setTimeZone(TimeZone.getTimeZone("UTC"));
-      return getGapInfoAsString(sdf);
-    }
-
-    public String getGapInfoAsString(DateFormat df) {
-      StringBuilder sb = new StringBuilder();
-      for (int j = 0; j < gapNameIdentifiers.length; ++j) {
-        sb.append(gapNameIdentifiers[j]);
-        sb.append(":\n");
-        for (int i = 0; i < gapStarts[j].length; ++i) {
-          sb.append("\t");
-          Date start = gapStarts[j][i];
-          Date end = gapEnds[j][i];
-          sb.append(df.format(start));
-          sb.append("\t");
-          sb.append(df.format(end));
-          sb.append("\n");
-        }
-        sb.append("\n");
-      }
-      return sb.toString();
-    }
-
-    public Date[][] getGapStartDates() {
-      return gapStarts;
-    }
-
-    public double[] getInitPoles() {
-      return numerMap.get("Initial_poles");
-    }
-
-    public double[] getInitZeros() {
-      return numerMap.get("Initial_zeros");
-    }
-
-    public byte[] getPhaseErrorImage() {
-      return imageMap.get("Phase_error");
-    }
-
-    public byte[] getPhaseImage() {
-      return imageMap.get("Response_phases");
-    }
   }
 
 }
